Improve the "versionreport" RPC output
This improves the usefulness of the "versionreport" RPC output by
aggregating version strings without the commit ID suffix. It adds
a configurable parameter to specify the number of blocks to count
versions backward from, structures output as pure JSON, and looks
back one day's blocks by default instead of 100 blocks as before.

The changes decouple the RPC method from the old quorum caches as
a step to prepare the application for an upcoming transition away
from the quorum system.