Port of the Bitcoin Logger to Gridcoin
This is the initial commit to port Bitcoin's Logger class to
Gridcoin. Templating has been used to handle the legacy string categories.
Mapping from the old fDebugxx flags to the the category enum is still in
progress.

Also remaining is the adding of the scraper logger style log rotation and archiving.