Decouple "network" RPC statistics from tally
The "network" RPC output shows data based on two-week aggregates of
payment amounts. Before, this function pulled the data via the two-
week averages periodically collected by the tally. The legacy tally
recount routine, however, counts two weeks using the BLOCKS_PER_DAY
constant that does not accurately reflect the reality of variations
in the number of blocks generated by the network over two weeks. It
also begins counting backward from an offset, so the statistics are
never current for the chain tip.

This replaces the use of the tally data for calculating the network
statistics with a direct scan of the blockchain and removes code in
the tally that exported those statistics before. The changes update
the output with standard money formatting and JSON field names.