Remove unneeded RA init messages to fix Qt segfault on Windows
Improvements to the tally system speed up the scan for research reward
blocks on boot. The wallet enqueues messages that provide feedback for
users about the number of blocks scanned so far. The speed improvement
causes too many of these messages to queue up before the splash screen
event loop can wake up to process those waiting.

This removes the feedback messages to avoid an overflow condition. The
RA scan executes so quickly anyway that users may never see the status
in the splash screen, so we can reduce the overhead a bit.