Track first reward block in research account
This tracks the first reward block of a research account instead of just
the timestamp.

It also fixes the ability to reorganize backward for those accounts with
magnitudes of zero for the first reward block. This problem is caused by
a questionable rule that skips the increment of the account accuracy for
a zero-magnitude research reward block.