Remove redundant LoadAdminMessages() calls
This removes unnecessary calls to LoadAdminMessages() to process the
contracts in new blocks. While not reorganizing, this function scans
back six blocks and reloads any contracts it finds. The offset isn't
necessary because we can load the contracts from transactions in the
connected blocks directly which avoids reading blocks from the disk.