Improve safety of public superblock fields
The m_height and m_timestamp fields of the Superblock class were
memory-only values that needed to be properly set when receiving
a new superblock. This was problematic because there was no good
place to copy the correct values so we risked using objects with
these fields uninitialized.

With the addition of the Quorum API, we used smart pointers that
allowed for shared ownership of Superblock objects. This extends
that pattern by adding a dedicated smart pointer type that wraps
a superblock and stores this context alongside it.