Add reasonable ETTS test to diagnostics
If the wallet is configured as a researcher wallet, this test
checks the ETTS. It uses the ETTS function with the parameter set
to ignoring whether the wallet is actively staking or not.

If the result is > 90 days, then the test is failed (red).
If the result is 45 days < result <= 90 days, then the test is warning (yellow).
If the result is <= 45 days, then the test is passed (green).

The idea behind this is to alert people who have decided to solo
mine but have insufficient balance to ensure that they stake before
the 180 limit expires.