Improve Staking Icon
Improvements have been made to the miner code to use scoped enums
for the miner status rather than a concatenated string. (The string
is also retained for legacy purposes for now.)

The old staking icon/tooltip conflated two different states together:
"unable to stake" because the wallet is locked or the reserve is set
so that no coins can be used for staking, etc. vs. "currently not staking,
but able to stake", which is all of the coins currently on cooldown
(such as for a new wallet that just received it's first "deposit", or a
one UTXO wallet where the UTXO just staked and is on cooldown.

I have introduced a new staking icon, a red down arrow, which indicates
"unable to stake", with an appropriate tooltip description. The gray
down arrow is now used for "currently not staking" status, where staking
will occur without any further operator action (i.e. coins maturing).
The green up arrow is used for currently staking status as before.

Some minor improvements were also made, such as minor style cleanups
of the code where I was working, along with some lock optimizations.