Convert scraper logging from fDebug3 to new BCLog::LogFlags::SCRAPER
This commit removes the fDebug3 conditionals from the scraper logging.
The use of the category is internal to the _log function, and is
used to control the posting of INFO level log entries only. Critical,
error, and warning level messages will still be logged in both logs
whether the SCRAPER category is on or off.