Fix beacon renewal for scraper consenus window
An issue observed on testnet resulted in exclusion of a CPID from a
superblock when that wallet automatically renewed its beacon during
the superblock formation period. The new advertisement replaced the
beacon with a timestamp later than the beacon consensus window used
by the scrapers to create a list of CPIDs to include for statistics
processing.

This change disables beacon renewal immediately prior to superblock
formation to avoid the condition above.