Replace boost::optional<T&> with non-owning pointers
I used boost::optional in a few places to return an optional type that
wraps a reference type. Although Boost supports this, I read that this
is not a good practice, and that the std::optional type in C++17 won't
allow a reference type for the template argument. This change replaces
the optional references with non-owning pointers in a way that retains
the public API.