Refactor claim context objects into contracts
This implements the IContractPayload interface for claim objects so that
we can remove the claim fields from blocks. Instead, we store a claim as
a contract in the coinbase transaction.

By doing so, we avoid the need to add and check the claim hash in legacy
hashBoinc fields to protect the integrity of the claim data since claims
are not part of a block hash.