Add preliminary testing option for block v11 height on testnet
This allows for configuration of the testnet block version 11 transition
height for early testing of hard fork changes. Specify -v11height=X from
the command line or add the directive to the configuration file to start
a node that will switch to the block version 11 protocol after the chain
grows to X number of blocks.