Fix depends bzip2 recipe for MinGW cross-compilation
This removes the bzip2 executable programs from the recipe. For MinGW
platforms, the Makefile attempts to copy executables with ".exe" file
extensions and bails when it cannot access files without the extenion
present. We only need the bzip2 libraries for the depends build path,
so we can ignore the executables.