Add Tally API for direct accrual calculation
This provides a method that calculates the accrual for a CPID without
the need to obtain an IAccrualComputer object beforehand. It improves
readability of client code that does not need those extra details and
may allow the compiler to perform some additional optimizations since
the virtual classes exist within the same translation unit.