Improve local BOINC project to whitelist matching
This fixes the ability for the wallet to match local BOINC projects to
the corresponding whitelisted projects when the project contracts have
names that do not match the canonical names stored by the BOINC client
in the client_state.xml file. The wallet will match projects by URL in
case an exact name match fails. The URLs in the project contracts will
usually contain the correct values even though the names may differ.