Optimize and refactor explain magnitude
After the removal of the the legacy "neural network" messaging system,
we no longer need to retain the old string format for explainmagnitude
so we can replace it with structured data. This greatly improves speed
and provides a usable API for integration with the GUI and other parts
of the application.