Add context for when BOINC is attached to a pool
This adds detection for Gridcoin pool accounts when loading the
researcher context to select a CPID. It allows the UI to supply
more useful information for pool users.

For now, this embeds pool CPIDs in the wallet. We may expand on
this in the future to allow pool operators to submit a contract
to register a pool through the blockchain.