Replace Contract::m_tx_timestamp field with tx context
This removes the m_tx_timestamp field from contract objects. The field
was a bit of a hack to pass around the transaction timestamp for those
objects. Instead, contract handlers now require transaction references
so this change uses the nTime field from the supplied context. Because
of this, the safety of the contracts API improves since we do not need
to remember to properly initialize the m_tx_timestamp fields after the
contract constructor.