Automate beacon advertisement for renewals only
This removes the automatic beacon advertisement for new users. The
wallet will only send an automatic transaction to renew beacons. A
new user is often confused by the initial onboarding process. When
the wallet sends an initial beacon without user action, it adds to
the confusion by masking a step in the documented process. The new
beacon wizard walks a user through this process so the wallet does
not need to send the beacon automatically.