Implement 1E-8 relative tolerance in CheckReward()
While Fern code goes to great lengths to remove sources
of rounding error from the research rewards accrual, some
double math remains, and can in certain circumstances
generate a very small amount of error in the calculations
of the rewards between the staking machine and the calculating
machine. This commit allows the claimed reward to exceed the
computed award at a rate of 1 Halford per 1 GRC of the total
computed award, which is a relative error of 1E-8.