Add transactionUpdated signal
With the hide loop in the updatedTransactions method in the
overview page, the updateTransactions needs to be called after
the underlying transaction model has recevied a new row, because
the getLimit is set to a higher value than is visible on purpose
to avoid a resort.

The signal that would normally be used, numTransactionsChanged,
does NOT work for this purpose, because that signal is only sent
when the size of the number of transactions in the wallet is changed.
The size will change if the wallet is just keys and a resync or
rescan is done, but the situation when an old wallet is used with
existing transactions, but the blockchain is synced from zero will
not work. The wallet size is constant then for old transactions
relinked.

To handle this case I had to add a new signal transactionUpdated
which is sent when WalletModel::updateTransaction is called.