Ensure contract validitiy before including a tx in a new block
This adds a check to the miner that performs contextual validation of
any contracts in a mempool transaction before including the contracts
in a block. It avoids a condition that may occur when another staking
node generates a block with a transaction that invalidates a contract
already accepted to the mempool. A node that includes the transaction
with the now-invalid contract will fail to produce a subsequent block
that the network accepts.