Fix ability to send contracts after GUI wallet unlock prompt
When the wallet is locked for staking only, contract transactions sent
following a GUI prompt for the wallet passphrase fail to execute since
the GUI's temporary unlock mechanism for the dialog doesn't modify the
fWalletUnlockStakingOnly flag. Checking this flag is outside the scope
of the SendContract() functions so this change removes that condition.