Switch Superblock::CpidIndex::TotalMagnitude() to return integers
This method returned a floating-point number. This switches the return
value to an integer. The upcoming changes to the voting system rely on
the total magnitude to calculate voting weight. The use of the integer
return type avoids floating-point discrepancies and allows the network
to potentially integrate the voting system into the protocol in future
enhancements.

Note that this change truncates the fractional portion of the network-
wide magnitude from the return value. It is not significant to current
voting weight calculations.