Remove useless "Detach databases at shutdown"
The "Detach databases at shutdown" UI checkbox doesn't do
anything, and hasn't in a while, and the state of the checkbox
does not even persist in the optionsmodel.

This has to be a throwback to when the blockchain index/transactions
were in bdb. Those have been in leveldb for a long while now,
and the wallet is always flushed and an LSN_reset (essentially
a detach) is issued upon successful shutdown of the wallet.