Adjust GetGeneratedType, decomposeTransaction and others
This commit adjusts GetGeneratedType and decomposeTransaction to
properly accomodate sidestake sends. It also retires the original
ListTransactions and GetAmounts, corrects ListTransactions2 and
GetAmounts2 and renames them to ListTransactions and GetAmounts
to properly accomodate sidestake send display, and also corrects
other aspects of ListTransactions and GetAmounts to correct
apparent balance mismatch between getbalance and getbalance "*" 0,
and also the output of listtransactions and gettransaction.