Implements a transactiontablemodel date filter
This commit provides a new check box in options/Display:

"Only display txns on or after". If checked, then there
is a date entry field. This will cause the transactiontablemodel
to only load transactions with a data on or after that date
from the core wallet.

This is very helpful when dealing with wallets that have huge
numbers of transactions as setting the date to a reasonably
recent past will reduce the size of the QList and restore
UI performance.

Note that this is different than the date filter on the transaction
table. That filter operates on the QList at the view level and
selecting something like "last week" will not help the UI pauses
for large number of transaction wallets, because the problem is
at the transactiontablemodel level.

If this date field is used, it limits what can be seen on the
transaction table view, although the date filters on the view
can be used to further refine if desired.

Note that this checkbox and date can be changed and the wallet
will update immediately when the apply and/or ok is clicked.