Prepare to drop legacy fields from version message
Early in its life, Gridcoin added some fields to CNode for the legacy
research reward system. The design of the code that used these fields
had security and privacy issues so later changes set the fields to an
empty string. This commit prepares nodes for the removal of fields in
legacy "version" messages in the release that will follow the pending
mandatory version.