Fix GUI signals for wallets started in investor mode
This fixes the GUI researcher context signals for a wallet that starts
in investor mode. In this mode, the GUI skipped the step that connects
those signals. Before the researcher wizard supported investor or pool
mode, the wallet required a restart to change from investor mode so it
did not need those signals. The wizard enables selection of every mode
at runtime so we must always handle the signals.