Implement scraper global statistics cache optimization
This commit changes the ConvergedManifestPartsMap to a
map of pointers, ConvergedManifestPartPtrsMap.

The CScraperManifest held bythe ConvergedManifest is via a shared_ptr.
In turn the CParts in the CScraperManifest::CSplitBlob vParts are also
mapped intothe ConvergedManifestPartPtrsMap, which is indexed by project.
The part pointers in the ConvergedManifestPartPtrsMap are valid
for the same lifetime as the CScraperManifest held by shared pointer
in the ConvergedManifest, so this is safe.

Thjs minimizes the changes to the rest of the scraper to achieve
this optimization. Some adjustments to the testing harness was
necessary. I also found and put in a missing IsComplete() check
in BinCScraperManifestsByScraper(), which may have been responsible
for the extremely rare occurence of missing projects in a local
node convergence and explainmagnitude.