Fix GUI OP_RETURN filter to avoid hiding transactions with messages
This changes the criteria for the filter that hides transactions that
contain OP_RETURN outputs to avoid squelching transactions that occur
after the v4.0.0 hard-fork. The filter fixes an old issue that caused
a wallet to display transactions that it did not send itself if those
transactions contain OP_RETURN outputs. Since contracts in version 2+
transactions use OP_RETURN outputs to burn the contract fee, we added
an additional constraint on the filter that limits it to tx version 1
only. As a side effect, the constraint hides legacy transactions with
message contracts during the transition period to block version 11.

I adjusted the constraint to filter by timestamps instead of version,
so the filter will only check transactions that occur before the time
of the previous mandatory release hard-fork. By this time, every node
ran a version of the wallet that ignores a transaction with OP_RETURN
outputs properly.