Fix blockchain stall and priority queue state
This reverts some changes added for scraper messages that sometimes
cause a node to skip sending a getdata message for blocks. This can
stall the node because it does not know about the parent block that
subsequent blocks build on.

I also uncommented a line needed for maintaining the priority queue
for inventory messages for a peer. This reduces the overhead of the
queue processor for getdata message.