Implement std::atomic_bool OutOfSyncByAge
This commit replaces the function call OutOfSyncByAge() with
a void UpdateOutOfSyncByAge() that updates an atomic boolean
fOutOfSyncByAge and an atomic int64_t nSyncTime. This function
is called in the "main loop" in init and main in the block processing
functions where a lock on cs_main is essentially already taken
at the only points where pindexBest is updated.

This allows us to change the old OutOfSyncByAge() function into
simply the fOutOfSyncByAge atomic boolean flag itself, which
requires no cs_main lock. This eliminates a possible source of
deadlocks, since a number of threads use OutOfSyncByAge and also
take other locks.