Fix regression for superblock builder optimization
This fixes a regression in the superblock builder that optimizes
construction of a superblock from scraper statistics. The change
that added the beacon IDs to the superblock altered the loop for
the scraper stats import in a way that reverted the optimization
that skips the unneeded byCPIDbyProject entries (the bulk of the
data structure).