Adjust threading for Gridcoin services
GRC::Initialize must be called as part of StartNode(), which
creates a thread group netThreads forked from a thread in the global
threads group in init. The scraper components are run in the
netThread groups as part of this, since they share data structures
with scraper_net, and the collapse of the threads has to be
coordinated.