Reimplement block index sanity check as a background job
This adds a start-up job that checks the block index for entries that
contain a null pointer to the previous block. With the removal of the
stake modifier checksum code, we reimplement this heuristic which can
identify corrupted block index disk state potentially caused by power
failures or unclean shutdowns.