Allocate block index researcher context only for research blocks
The Gridcoin researcher context in the block index contained these
fields:

 - cpid (16 bytes)
 - nResearchSubsidy (8 bytes)
 - nMagnitude (8 bytes)

However, non-researcher nodes produce the overwhelming majority of
Gridcoin's blockchain. This change allocates memory for researcher
context objects only for blocks staked with a CPID. We add 8 bytes
for pointers for every block but save 24 bytes for all of the non-
CPID blocks that comprise nearly three-quarters of the chain.