Enhance ETTS calculation
This commit implements a two step process for computing the
underlying difficulty used for the GetEstimatedTimetoStake function.

This is meant to address the matching of the appropriate span to
use for the difficulty calculation compared to the ETTS (which is
a coupled problem). For low balance holders, a long (960 block)
averaging interval should be used to compute the difficulty
for estimating ETTS, since if someone has an ETTS of 1 month, it
makes no sense to be changing that estimate based on difficulty
changes during the last hour. Conversely, if someone is staking
6 times a day, a much shorter span for the difficulty is appropriate
since changes in the network difficulty will be reflected very
quickly with staking frequency.