Port Bitcoin intro.h/cpp/ui
This commit ports the Bicoin intro.h, .cpp,and .ui files,
which contain the implementation of the GUI Intro class. This
provides the implementation support of the intro ui form (dialog
box), which allows the user to specify a data directory on a
brand new installation of Gridcoin and have it remembered in the
options model (Qt GUI config file).