Implement MilliTimer class and add basic timing logging to StakeMiner
This commit implements a nifty little timer class that can be used
as a local timer or instantiated as a global timer object with a
map of timers.

A single global timer instance g_timer is initialized in the timer.cpp
file with the "default" label. The StakeMiner timings use a new "miner"
label in the global timer to allow timings that span across mulitple
functions/methods.