Implement staking loop efficiency measurement
This adjusts the miner to count the number of staking intervals
that saw one or more successful UTXO mining loops. A measure
of staking loop efficiency is then calculated by dividing that
counter by the total number of intervals that have elapse since
wallet start up. The is a good measure of staking coverage.