Implement staking efficiency measure
This commit refines the calculation of staking_loop_efficiency,
which measures the percent of time that the staking loop is staking.

It also adds the measure of staking_efficiency, which includes the
combined effect of the quantized (masked) intervals covered by
the miner, and also the actual weight applied in the staking loop
versus the ideal weight. Note that this includes the effect of
cooldown and UTXO's that are too small to stake.