Refactor SelectCoinsForStaking() and AvailableCoinsForStaking()
This commit refactors the above two calls to remove the redundant
iterations through the entire wallet map and also unnecessary
iterations through the Block map.

It also removes unnecessary transaction and block file lookups
in the miner's CreateCoinStake in favor of using the mapBlockIndex
in memory, and does some other minor cleanup.