Add reset button to clear coincontrol entries
When advanced coin control is enabled, and via the inputs button
inputs are selected, then those inputs are spent via an underlying
transaction or staking, they become "stuck". This is because
the coinControl object still has the stale COutPuts in it.

A simple reset button has been added that calls SetNull() to
nullify the coinControl object and then updates the coin control
labels. This gets rid of the stuck selected inputs.