Implement CAmount GetNewbieSuperblockAccrualCorrection(Cpid cpid)
This commit implements the accrual correction function necessary to
fix the newbie accrual bug and also implements changes to the auditing
functions. Note that a historical beacon map that is effectively a
linked list is also implemented and the lookback scope of both the
audit and any accrual correction is limited to the chain of beacon
renewals going back to the beacon advertisement from the current active
beacon (if it is a renewal).

It also updates the testnet newbie fix fork point to 1480000 and
mainnet to std::numeric_limits<int>::max() for right now.