Tweak BeaconRegistry::Revert
The Revert function was identifying v1 direct to active beacons
to revert as renewals rather than direct activations (i.e. before
the v1 ADD contract itself was reached. This is innocuous, because
in reality we don't care about the beacon chains before the Farn
hard fork height anyway. But for correctness this is fixed.