Add capability to close researcher repository
This adds the capability to close the researcher registry file
to remove the lock on that file so that the snapshot cleanup will
succeed when using the snapshot download functionality.

It also changes the Qt snapshot upgrader class to use the existing
QApplication app instead of a new one, because earlier to
support the data directory dialog box the place where QApplication app
was instantiated was changed to much earlier.

Having another QApplication on top of the existing one was causing
a segfault at exit.