Update Qt depends recipe from Bitcoin
bitcoin/bitcoin@1112035d32ffe73a4522226c8cb2f6a5878d3ada

This updates to the latest Qt depends recipe from Bitcoin. It supports
fixes for build problems for the GUI on macOS. As a small side effect,
this upgrades the depends Qt version from 5.9.6 to 5.9.8.

Note that Qt 5.9 is not compatible with the macOS dark-mode feature. A
separate effort will upgrade the recipe to 5.12 or later to avoid some
display issues with Gridcoin's "native" theme on a dark-mode desktop.

This change also updates the Qt Charts and SVG modules of the original
Qt recipe for Gridcoin and removes the now unused ActiveQt module that
provided COM support for the legacy "neural network" on Windows.