Fix depends Qt path resolution for auxillary modules
This adds a top-level project to the Qt depends recipe that enables
the build system to generate the correct pkgconfig paths for all of
the platform targets.

Gridcoin uses the Charts and SVG modules in additon to the basic Qt
UI components needed upstream by Bitcoin. Some platform tool chains
ignored the invalid pkgconfig paths generated for these modules. To
fix the static Qt builds for every target, we need to address those
issues because the previous work-around for one broke another.