Make some minor adjustments for VotingDialog flow
This commit allows going to the voting dialog page without
automatically kicking off calculating the poll results to avoid
the long term lock every time the voting page is selected.

It also implements a one hour timer once the data is refreshed,
and will hide the polling data asking the user to refresh it
once the timer is expired.