Port install_db4.sh tool from Bitcoin, add to macOS CI
The install_db4.sh tool is useful for obtaining the 4.8 version of
the Berkeley DB library needed for the wallet database. We add the
script here specifically to support macOS CI for GitHub Actions. A
quirk exists between the Xcode environment on the platform and the
berkeley-db4 Homebrew formula that causes the wallet tests to fail
when linked against Homebrew's recent build of BDB 4.8.

Instead of relying on the formula, we call the install_db4.sh tool
to install the dependency. This will build Berkeley DB without the
replication feature that exhibits this problem.

In the future, we may be able to return to installing the Homebrew
package to save time if the formula resolves the issue.