Deal with Qt depends .qmake.stash file
Gridcoin added a top-level Qt SUBDIRS-type project in #2038 (6f090a0)
to manage auxillary Qt modules needed by the project, and this causes
Qt to generate a .qmake.stash file in the depends directory. The file
contains a cache of qmake build variables which can cause issues with
subsequent static builds that target a different platform or SDK than
those selected for the previous depends build.

This ensures that the Qt recipe clears the stash file when rebuilding
the Qt packages, and qmake will regenerate it as needed.