Fix another subtle error in renewal chain walker
In GetNewbieSuperblockAccrualCorrection the while loop that walks
the beacon chain backwards to the original advertisement has a
mistake and uses beacon instead of beacon_ptr in the find. As a
result it will not work for more than one renewal cycle.

This is not a consensus issue, because this function is not called
for beacons that have more than one renewal cycle. In fact, now
that the newbie bug is out of the way, it will not be called for
beacons with any renewal at all. So this is merely for correctness.