Reduce default connection limit back to 125
The default value for "-maxconnections" was increased from 125 to 250
back in 2014 without a documentation change. This decreases the limit
back to 125 connections for listening nodes. 250 connections is a bit
high as a default for unsuspecting node operators and it might create
an unexpected burden for a device with limited resources. Since a GUI
wallet is a listening node by default, we err on the side of caution.
Folks that wish to service a greater number of connections can change
the "maxconnections" option or set the directive in the configuration
file:

    maxconnections=250