Reimplement checkpoint height validation
Commit c949a02436374bc8b57d01a0195054f9397d0ccd removed broadcasted
checkpoints and reimplemented Checkpoints::CheckSync() for hardened
checkpoints. However, the change did not apply the validation again
for received blocks.

This reimplements the validation using a similar strategy for nodes
that finished the initial sync. With this change, nodes will refuse
blocks that connect to a height below the highest checkpoint.