Fix deferment of beacon renewal in superblock window
This fixes a condition that was supposed to prevent a beacon from
disappearing from a superblock for one cycle if the wallet renews
the beacon during the scraper activity period before a superblock
is due (21621a675f61d8b3dbabdf1999e8b9064b986d3a).

The original change didn't account for the additional time of the
scraper activity window before the superblock.