Fix GUI macOS and designer font sizes
This change enables an increased font size for Windows and Linux only.
Because macOS often uses a different reference DPI, a fixed point-size
adjustment like we defined before can actually result in smaller text.
Instead, we rely on the default font size for macOS builds.

This also fixes a side-effect of the global stylesheet font size value
which overrides the sizes configured for text in the designer forms.