Set explicit base font size for macOS
This fixes the rendered size for widgets that have a smaller or larger
font size defined by Qt's macOS style. In particular, it increases the
text size of QToolButton widgets in the sidebar to improve legibility.

Since macOS uses a different base OS-level DPI than Windows and Linux,
the applicaton needs a larger base point size when rendering for macOS
desktops.