Switch GUI overview page form layouts to grids
This improves the rendering of the tables on the overview page on macOS.
Qt draws macOS form layouts with right-aligned labels and moves the form
to the center of its parent which gives the overview page an awkward and
broken appearance. Using grid layouts avoids the custom styles and shows
the information as on the other platforms.